+++ rust-connectors/sinks/kafka/CHANGELOG.md
@@ -1,3 +1,6 @@
 # Kafka Sink Change Log
 
+## kafka-sink Version 0.2.1 - UNRELEASED
+* Added support to smartmodule-parameters and smartmodule without explicit type. ([PR #287](https://github.com/infinyon/fluvio-connectors/pull/287)
+
 ## kafka-sink Version 0.1.0 - UNRELEASED