+++ rust-connectors/sources/mqtt/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Connector Change Log
 
+## mqtt Version 0.3.1 - UNRELEASED
+* Added support to smartmodule-parameters and smartmodule without explicit type. ([PR #287](https://github.com/infinyon/fluvio-connectors/pull/287)
+
 ## mqtt Version 0.2.0 - UNRELEASED
 * Add TLS Support for MQTT Connector ([PR #179](https://github.com/infinyon/fluvio-connectors/pull/179))